Add SetSimplify for lossy simplification
The new SetSimplify() function will simplfy a geometry upon rendering.
This performs a lossy converversion of the geometry, which may drop
points, making the layer look different that the original, but causes
Render() to return fewer bytes. The value can be 0.0 - 1.0, where 0
is no simplication an 1.0 is the maximum. Default is 0.0

It's worth noting that Render(), on it's own, provides automatic
lossless simplification that is generally good enough at most zoom
levels.